Set is_page to true for media page
Media page need to be set to true for is_page to work on media page
Also added the filter 'rtmedia_is_page' to make is_page to false.
Setting is_page to true will be useful for the themes which uses the
customizer to make the page one or multicolumn.